FEAT: Simplify and speed up the in-order retain()
Successive improvements and tweaks of this code lead to the following
version, which seems more obvious from the start (but it wasn't
obvious).

Instead of splitting retain into one tombstone-marking phase 1 and
a backwards shifting phase 2, we simply backwards shift eagerly during
the first and now only phase.

Improves performance of `.retain()`. For benchmarks "retain_few" and
"retain_half" it is now faster than `.retain_unordered()`.