FIX: Reduce generic parameter use in apply to index
Use a free function instead of a method so that apply_to_index is not
multiply compiled per every Key, Value type. Yes, a free function nested
inside a method should work fine -- it does not use the outer type
parameters.

Also added a free function case in the size dispatch macro to help with
this.