FIX: self-referential usage of indexmap/indexset macros
This was just fixed in maplit, so copying the same fix to here.

The attribute option fixes the self-referential usage of the macro in
itself, and fixes the macros so they can be used with their full path
(like indexmap::indexmap!(), see tests).

This is backwards compatible, see
https://doc.rust-lang.org/edition-guide/rust-2018/macros/macro-changes.html#local-helper-macros
for more information.