Build serde without default features
We don't actually need any std/alloc features of `serde`, but our
dev-dependency on `serde_test` was indirectly turning on its default
features. However, if we isolate that testing to a separate crate in the
workspace, Cargo will compute the enabled features separately. Now we
can actually enable the `serde-1` feature in `test-nostd` too!